backend/requirements.txt
langchain==0.2.16
langchain-community==0.2.16
sentence-transformers==3.1.1
faiss-cpu==1.8.0

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1

pymupdf==1.24.10
python-multipart==0.0.9

httpx==0.27.2